+++ epan/to_str.h
@@ -94,11 +94,11 @@ gchar*	guid_to_str_buf(const e_guid_t*, gchar*, int);
 
 WS_DLL_PUBLIC char *decode_bits_in_field(const guint bit_offset, const gint no_of_bits, const guint64 value);
 
-WS_DLL_PUBLIC char	*other_decode_bitfield_value(char *buf, const guint32 val, const guint32 mask,
+WS_DLL_PUBLIC char	*other_decode_bitfield_value(char *buf, const guint64 val, const guint64 mask,
     const int width);
-WS_DLL_PUBLIC char	*decode_bitfield_value(char *buf, const guint32 val, const guint32 mask,
+WS_DLL_PUBLIC char	*decode_bitfield_value(char *buf, const guint64 val, const guint64 mask,
     const int width);
-WS_DLL_PUBLIC const char *decode_numeric_bitfield(const guint32 val, const guint32 mask, const int width,
+WS_DLL_PUBLIC const char *decode_numeric_bitfield(const guint64 val, const guint64 mask, const int width,
     const char *fmt) G_GNUC_PRINTF(4, 0);
 
 WS_DLL_PUBLIC const gchar* port_type_to_str (port_type type);
